Struct hrpc::proto::Error

pub struct Error {
    pub identifier: String,
    pub human_message: String,
    pub more_details: Bytes,
}

Error type that will be returned by servers.

Fields
identifier: String

The identifier of this error, can be used as an i18n key.

human_message: String

A (usually english) human message for this error.

more_details: Bytes

More details about this message. Is dependent on the endpoint.

Implementations

impl Error

pub fn new_not_implemented(message: impl Into<String>) -> Self

Create a new hRPC error representing a not implemented endpoint (HrpcErrorIdentifier::NotImplemented).

pub fn new_resource_exhausted(message: impl Into<String>) -> Self

Create a new hRPC error representing resource exhaustion by a client (HrpcErrorIdentifier::ResourceExhausted).

pub fn new_internal_server_error(message: impl Into<String>) -> Self

Create a new hRPC error representing an internal server error (HrpcErrorIdentifier::InternalServerError).

pub fn new_not_found(message: impl Into<String>) -> Self

Create a new hRPC error representing a not found error (HrpcErrorIdentifier::NotFound).

pub fn with_details(self, details: impl Into<Bytes>) -> Self

Set the “more details” of this hRPC error.

pub fn with_identifier(self, identifier: impl Into<String>) -> Self

Set the “identifier” of this hRPC error.

pub fn with_message(self, message: impl Into<String>) -> Self

Set the “human message” of this hRPC error.