Trait hrpc::server::error::CustomError

pub trait CustomError: Debug + Send + Sync + 'static {
    fn error_message(&self) -> Cow<'_, str> { ... }

    fn status(&self) -> StatusCode { ... }

    fn identifier(&self) -> Cow<'_, str> { ... }

    fn more_details(&self) -> Bytes { ... }

    fn as_hrpc_error(&self) -> HrpcError { ... }

    fn as_error_response(&self) -> HttpResponse { ... }
}

Trait that needs to be implemented to use an error type with a generated service server.

Provided methods

fn error_message(&self) -> Cow<'_, str>

Error message for humans.

fn status(&self) -> StatusCode

HTTP status code that will be used in client response.

fn identifier(&self) -> Cow<'_, str>

Error identifier.

fn more_details(&self) -> Bytes

More details for this error.

fn as_hrpc_error(&self) -> HrpcError

Create an hRPC protocol error from this custom error.

fn as_error_response(&self) -> HttpResponse

Create a response from this error.