Trait hrpc::server::Server

pub trait Server: Send + 'static {
    fn make_routes(&self) -> Routes;

    fn combine_with<Other, OtherSvc>(
        self, 
        other: Other
    ) -> ServerStack<Other, Self>
    where
        Other: Server,
        Self: Sized,
    { ... }

    fn into_make_service(self) -> IntoMakeService<Self>
    where
        Self: Sized,
    { ... }

    fn layer<L>(self, layer: L) -> LayeredServer<L, Self>
    where
        L: Layer<Handler, Service = Handler> + Send + 'static,
        Self: Sized,
    { ... }

    fn serve<'async_trait, A>(
        self, 
        address: A
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        A: ToSocketAddrs + Send,
        A::Iter: Send,
        Self: Sized,
        A: 'async_trait,
        Self: 'async_trait,
    { ... }
}

Required methods

fn make_routes(&self) -> Routes

Creates a Routes, which will be used to build a RoutesFinalized instance.

Provided methods

fn combine_with<Other, OtherSvc>(self, other: Other) -> ServerStack<Other, Self> where
Other: Server,
Self: Sized,

Combines this server with another server.

fn into_make_service(self) -> IntoMakeService<Self> where
Self: Sized,

Turns this server into a type that implements MakeService, so that it can be used with [hyper].

fn layer<L>(self, layer: L) -> LayeredServer<L, Self> where
L: Layer<Handler, Service = Handler> + Send + 'static,
Self: Sized,

Layers this server with a layer that transforms handlers.

fn serve<'async_trait, A>(
self,
address: A
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
A: ToSocketAddrs + Send,
A::Iter: Send,
Self: Sized,
A: 'async_trait,
Self: 'async_trait,

Serves this server. See utils::serve for more information.