Struct hrpc::Request

pub struct Request<T> { /* fields omitted */ }

A hRPC request.

Implementations

impl<T> Request<T>

pub fn header_map(&self) -> Option<&HeaderMap>

Get an immutable reference to the header map.

pub fn header_map_mut(&mut self) -> Option<&mut HeaderMap>

Get a mutable reference to the header map.

pub fn get_or_insert_header_map(&mut self) -> &mut HeaderMap

Get a mutable reference to the header map, inserting a new one if it doesn’t already exist.

pub fn http_extensions(&self) -> Option<&Extensions>

Get an immutable reference to the HTTP extensions.

pub fn http_extensions_mut(&mut self) -> Option<&mut Extensions>

Get a mutable reference to the HTTP extensions.

pub fn get_or_insert_http_extensions(&mut self) -> &mut Extensions

Get a mutable reference to the HTTP extensions, inserting a new one if it doesn’t already exist.

impl<T> Request<T>

pub fn http_method(&self) -> Option<&Method>

Get an immutable reference to the HTTP method.

pub fn http_version(&self) -> Option<&Version>

Get an immutable reference to the HTTP version.

pub fn http_uri(&self) -> Option<&Uri>

Get an immutable reference to the HTTP URI.

impl Request<()>

pub fn empty() -> Request<()>

Create a request with an empty body.

This is useful for hRPC socket requests, since they don’t send any messages.

impl<T> Request<T>

pub fn new_with_body(body: Body) -> Self

Creates a new request using the provided body.

pub fn extensions_mut(&mut self) -> &mut Extensions

Get a mutable reference to the extensions of this response.

pub fn extensions(&self) -> &Extensions

Get an immutable reference to the extensions of this response.

pub fn endpoint_mut(&mut self) -> &mut Cow<'static, str>

Get a mutable reference to the endpoint of this response.

pub fn endpoint(&self) -> &str

Get an immutable reference to the endpoint of this response.

impl<T: PbMsg> Request<T>

pub fn new(msg: &T) -> Self

Create a new request with the specified message.

impl<T: PbMsg + Default> Request<T>

pub async fn into_message(self) -> Result<T, DecodeBodyError>

Extract the body from the request and decode it into the message.