Crate hrpc

Common code used in hRPC code generation.

Modules

body

Body type.

client

Common client types and functions.

exports

Some re-exported crates that might be useful while writing software with hrpc.

server

Common server types and functions.

Macros

include_proto

Include generated proto server and client items.

return_error

Returns an error.

Structs

Request

A hRPC request.

Response

hRPC response type.

Enums

DecodeBodyError

Constants

HRPC_HEADER

The hRPC protobuf mimetype.

Traits

IntoRequest

Trait used for blanket impls on generated protobuf types.

IntoResponse

Trait used for converting any type to a Response type.

Functions

encode_protobuf_message

Encodes a protobuf message into a new BytesMut buffer.

encode_protobuf_message_to

Encodes a protobuf message into the given BytesMut buffer.

Type Definitions

BoxError

Alias for a type-erased error type.

HttpRequest

HTTP request used by hRPC.

HttpResponse

HTTP response used by hRPC.