Trait hrpc::server::CustomError

pub trait CustomError: Debug + Display + Send + Sync {
    fn code(&self) -> StatusCode;

    fn message(&self) -> Vec<u8>;

    fn decode_error() -> (StatusCode, Vec<u8>) { ... }

    fn not_found_error() -> (StatusCode, Vec<u8>) { ... }

    fn internal_server_error() -> (StatusCode, Vec<u8>) { ... }
}

Trait that needs to be implemented to use an error type with a generated service server.

Required methods

fn code(&self) -> StatusCode

Status code that will be used in client response.

fn message(&self) -> Vec<u8>

Message that will be used in client response.

Loading content...
Provided methods

fn decode_error() -> (StatusCode, Vec<u8>)

Return a message decode error.

fn not_found_error() -> (StatusCode, Vec<u8>)

Return a not fonud error.

fn internal_server_error() -> (StatusCode, Vec<u8>)

Return an internal server error.

Loading content...